// web/src/lib/schemas.ts
export const BudgetItemBase = z.object({
  actualAmount: z.number(),
  targetAmount: z.number(),
  budgetId: IdSchema,
  categoryId: IdSchema,
});

export const BudgetItemSchema = addIdAndDatetime(BudgetItemBase);
export type BudgetItem = z.infer<typeof BudgetItemSchema>;

// Category
export const CategoryBase = z.object({
  name: z.string(),
  group: z.string(),
  status: z.string(),
  userId: IdSchema,
});

export const CategorySchema = addIdAndDatetime(CategoryBase);
export type Category = z.infer<typeof CategorySchema>;